Add files via upload
Simple program intended to read a sentence, recognize if it has punctuation, and the output should be the words in reverse but with the letters in the original order. For example, the string “Dog bites man” should output as “man bites Dog.” Additionally first word is to be capitalized and punctuation is to remain at the end.